Move screens and navigation to ui package
