+++ app/src/main/java/com/example/listifyjetapp/MainActivity.kt
@@ -15,7 +15,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.listifyjetapp.navigation.ListifyNavigation
+import com.example.listifyjetapp.ui.navigation.ListifyNavigation
 import com.example.listifyjetapp.ui.theme.ListifyJetAppTheme
 import dagger.hilt.android.AndroidEntryPoint
