+++ app/src/main/java/com/example/listifyjetapp/ui/navigation/ListifyNavigation.kt
@@ -1,11 +1,11 @@
-package com.example.listifyjetapp.navigation
+package com.example.listifyjetapp.ui.navigation
 
 import androidx.compose.runtime.Composable
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.example.listifyjetapp.screens.lists.ListifyListsScreen
-import com.example.listifyjetapp.screens.splash.ListifySplashScreen
+import com.example.listifyjetapp.ui.screens.lists.ListifyListsScreen
+import com.example.listifyjetapp.ui.screens.splash.ListifySplashScreen
 
 
 @Composable
